move common modules to util file
