+++ packages/flex-plugin-scripts/src/scripts/start.ts
@@ -17,14 +17,7 @@ import {
   startIPCServer,
 } from '../config/devServer/ipcServer';
 import webpackDevServer from '../config/devServer/webpackDevServer';
-
-const PLUGIN_INPUT_PARSER_REGEX = /([\w-]+)(?:@(\S+))?/;
-
-interface UserInputPlugin {
-  name: string;
-  remote: boolean;
-  version?: string;
-}
+import { findFirstLocalPlugin, parseUserInputPlugins, UserInputPlugin } from '../utils/parser';
 
 interface StartServerOptions {
   port: number;
@@ -67,9 +60,8 @@ export const start = async (...args: string[]): Promise<StartScript> => {
   // Future  node version will silently consume unhandled exception
   process.on('unhandledRejection', err => { throw err; });
 
-  const userInputPlugins = _parseUserInputPlugins(...args);
-  const localPlugin = userInputPlugins.find(p => !p.remote);
-  const plugin = localPlugin && readPluginsJson().plugins.find((p) => p.name === localPlugin.name);
+  const userInputPlugins = parseUserInputPlugins(...args);
+  const plugin = findFirstLocalPlugin(userInputPlugins);
   if (!plugin) {
     throw new FlexPluginError('You must run at least one plugin locally.');
   }
@@ -214,44 +206,6 @@ export const _updatePluginPort = (port: number, name: string) => {
   });
 
   writeJSONFile(getPaths().cli.pluginsJsonPath, config);
-};
-
-/**
- * Reads user input to returns the --name plugins
- * --name plugin-test will run plugin-test locally
- * --name plugin-test@remote will run plugin-test remotely
- * --include-remote will include all remote plugins
- * @param args
- */
-export const _parseUserInputPlugins = (...args: string[]): UserInputPlugin[] => {
-  const userInputPlugins: UserInputPlugin[] = [];
-  const config = readPluginsJson();
-
-  for (let i = 0; i < args.length - 1; i++) {
-    if (args[i] !== '--name') {
-      continue;
-    }
-    const groups = args[i + 1].match(PLUGIN_INPUT_PARSER_REGEX);
-    if (!groups) {
-      throw new Error('Unexpected plugin name format was provided');
-    }
-
-    const name = groups[1];
-    const version = groups[2]; // later we'll use this for the @1.2.3 use case as well
-
-    if (version === 'remote') {
-      userInputPlugins.push({name, remote: true});
-      continue;
-    }
-
-    const plugin = config.plugins.find((p) => p.name === name);
-    if (!plugin) {
-      throw new FlexPluginError(`No plugin file was found with the name \'${name}\'`);
-    }
-    userInputPlugins.push({name: plugin.name, remote: false});
-  }
-
-  return userInputPlugins;
 };
 
 run(start);
